- Change const return data type for error values from int to size_t. - Removed get(TX _x).

+++ PrecompData_base.h
@@ -32,13 +32,13 @@ public:
 
 public:
 	// Error return values
-	static const int err_no_data              = -1,
-	                 err_device_not_available = -2;
+	static const size_t err_no_data              = -1,
+	                    err_device_not_available = -2;
 
 	// Warning return values
-	static const int wrn_x_less_than_min      = -101,
-	                 wrn_x_more_than_max      = -102,
-	                 wrn_invalid_oversampling = -103;
+	static const size_t wrn_x_less_than_min      = -101,
+	                    wrn_x_more_than_max      = -102,
+	                    wrn_invalid_oversampling = -103;
 
 public:
 	// Test
